[Refactor] Remove PreloadLink
