+++ src/Apps/Artist/ArtistApp.tsx
@@ -5,12 +5,9 @@ import { AppContainer } from "Apps/Components/AppContainer"
 import { HorizontalPadding } from "Apps/Components/HorizontalPadding"
 import { track } from "Artsy/Analytics"
 import * as Schema from "Artsy/Analytics/Schema"
-import { PreloadLinkState } from "Artsy/Router/Components/PreloadLink"
-import { LoadingArea } from "Components/v2/LoadingArea"
 import React from "react"
 import { LazyLoadComponent } from "react-lazy-load-image-component"
 import { createFragmentContainer, graphql } from "react-relay"
-import { Subscribe } from "unstated"
 import { ArtistHeaderFragmentContainer as ArtistHeader } from "./Components/ArtistHeader"
 
 import {
@@ -49,21 +46,9 @@ export class ArtistApp extends React.Component<ArtistAppProps> {
           <Row>
             <Col>
               <NavigationTabs artist={artist} />
-
               <Spacer mb={3} />
 
-              {/*
-              When clicking nav links, wait for fetch to complete before
-              transitioning to new route
-            */}
-
-              <Subscribe to={[PreloadLinkState]}>
-                {({ state: { isLoading } }: PreloadLinkState) => {
-                  return (
-                    <LoadingArea isLoading={isLoading}>{children}</LoadingArea>
-                  )
-                }}
-              </Subscribe>
+              {children}
             </Col>
           </Row>
