+++ src/Apps/Collect/__tests__/routes.test.tsx
@@ -1,12 +1,9 @@
-// import { mount } from "enzyme"
+import { Boot } from "Artsy/Router/Boot"
 import { createMockNetworkLayer, renderUntil } from "DevTools"
 import { Resolver } from "found-relay"
-// import createRender from "found/lib/createRender"
-// import getFarceResult from "found/lib/server/getFarceResult"
 import { find } from "lodash"
 import React from "react"
 import { Environment, RecordSource, Store } from "relay-runtime"
-import { Boot } from "../../../Artsy/Router/Components/Boot"
 import { routes } from "../routes"
 
 describe("Routes", () => {
